+++ CHANGELOG.md
@@ -53,6 +53,7 @@ All notable changes to this project will be documented in this file.
 - AS1/2 - Infinity, NaN can be used as identifiers, are normal variables
 - AS2 - obfuscated class attribute names
 - AS1/2 - newobject deobfuscated name
+- AS2 - obfuscated extends, implements
 
 ## [14.6.0] - 2021-11-22
 ### Added